Task 4 : Type II error
In our case, the null hypothesis is that the three samples are not statistically different. And we know that they are, in fact, statistically different because we "mocked up" the data. However, if we run the experiment many many times, we might expect to see that the means of the 3 samples are the same in a very, very small number of cases due to randomness. In such an unlikely case, we would incorrectly conclude that the samples are not statistically different and fail to reject the null hypothesis. This is a type II error.